backend/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.7.0
attrs==24.3.0
click==8.1.8
discord.py==2.4.0
fastapi==0.115.6
frozenlist==1.5.0
h11==0.14.0
idna==3.10
multidict==6.1.0
pika==1.3.2
propcache==0.2.1
pydantic==2.10.4
pydantic_core==2.27.2
python-dotenv==1.0.1
python-multipart==0.0.20
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.34.0
yarl==1.18.3
